渲染代码调试

+++ core/easy2d/src/gui/e2d_text_field.h
@@ -31,8 +31,8 @@ namespace Easy2D
 
         virtual Vec2 getDimensions() const;
 
-        int32 getWidth() const;
-        int32 getHeight() const;
+        float32 getWidth() const;
+        float32 getHeight() const;
 
     protected:
         SPtr<TextComponent> mTextComponent;
